Move BlogNav into its own component
