+++ styles/blog.module.css
@@ -3,31 +3,6 @@
   margin: 0 auto;
 }
 
-.blogNavigation {
-  display: flex;
-  justify-content: space-between;
-  gap: 2em;
-}
-
-.blogNavigation a {
-  text-align: center;
-  text-decoration: none;
-  color: var(--color-text-accent);
-  font-style: italic;
-  padding: 0.05em;
-}
-
-.blogNavigation a:hover {
-  box-shadow: inset 0 -0.1em 0 0 var(--color-text-accent);
-}
-
-.blogNavigationSeparator {
-  display: none;
-  margin: 2px;
-  background: var(--color-text-accent);
-  width: 1px;
-}
-
 .blogPostHeading {
   font-family: 'Playfair Display', serif;
   font-size: 1.5rem;
@@ -59,10 +34,6 @@
 @media (max-width: 1024px) {
   .blogPage {
     width: auto;
-  }
-
-  .blogNavigationSeparator {
-    display: flex;
   }
 
   .blogHeaders {
